1. update wechaty grpc to 0.16.1 2. move examples from wechaty to independent module. Make example easy to use.
