+++ README.md
@@ -227,19 +227,20 @@ If you are interested in the translation and want to look at how it works, it wi
 
 ## Usage
 
-1. Add a token to class `io.github.wechaty.examples.Main` in folder `wechaty/src/main/java`
+1. Add a token to class `io.github.wechaty.example.Main` in folder `Examples/src/main/java`
 2. Build:
 
     ```shell
+    cd examples/  
     mvn install
     ```
 
 3. Run
 
     ```shell
-    java -jar wechaty/target/wechaty-1.0.0-SNAPSOHOT-jar-with-dependencies.jar   
+    java -jar target/examples-0.1.1-SNAPSHOT-jar-with-dependencies.jar 
     # or run in background  
-    nohup java -jar wechaty/target/wechaty-1.0.0-SNAPSOHOT-jar-with-dependencies.jar &>> nohup.out & tailf nohup.out
+    nohup java -jar target/examples-0.1.1-SNAPSHOT-jar-with-dependencies.jar &>> nohup.out & tailf nohup.out
     ```
 
 4. enjoy
@@ -262,6 +263,10 @@ mvn install wechaty
 ## History
 
 ### master
+
+### v0.1.1 (June 6 2020)
+1. update wechaty grpc to 0.16.1
+2. move examples from wechaty to independent module. Make example easy to use.
 
 ### v0.1.1 (May 31 2020)
