+++ pom.xml
@@ -37,7 +37,7 @@
 
             <dependency>
                 <groupId>io.github.wechaty</groupId>
-                <version>0.11.25</version>
+                <version>0.16.1</version>
                 <artifactId>grpc</artifactId>
             </dependency>
 
@@ -140,28 +140,12 @@
                 <version>2.6</version>
             </dependency>
 
-
-<!--            <dependency>-->
-<!--                <groupId>com.tinder.statemachine</groupId>-->
-<!--                <artifactId>statemachine</artifactId>-->
-<!--                <version>0.2.0</version>-->
-<!--            </dependency>-->
-
             <!-- https://mvnrepository.com/artifact/com.squareup.okhttp3/okhttp -->
             <dependency>
                 <groupId>com.squareup.okhttp3</groupId>
                 <artifactId>okhttp</artifactId>
                 <version>4.6.0</version>
             </dependency>
-
-
-<!--            &lt;!&ndash; https://mvnrepository.com/artifact/com.lmax/disruptor &ndash;&gt;-->
-<!--            <dependency>-->
-<!--                <groupId>com.lmax</groupId>-->
-<!--                <artifactId>disruptor</artifactId>-->
-<!--                <version>3.4.2</version>-->
-<!--            </dependency>-->
-
 <!--            &lt;!&ndash; https://mvnrepository.com/artifact/io.grpc/grpc-kotlin-stub &ndash;&gt;-->
 <!--            <dependency>-->
 <!--                <groupId>io.grpc</groupId>-->
