+++ wechaty-puppet/src/main/kotlin/Puppet.kt
@@ -12,7 +12,6 @@ import io.github.wechaty.io.github.wechaty.watchdag.WatchDog
 import io.github.wechaty.io.github.wechaty.watchdag.WatchdogFood
 import io.github.wechaty.io.github.wechaty.watchdag.WatchdogListener
 import io.github.wechaty.listener.*
-import io.github.wechaty.memorycard.MemoryCard
 import io.github.wechaty.schemas.*
 import io.github.wechaty.utils.FutureUtils
 import io.github.wechaty.utils.JsonUtils
@@ -42,7 +41,7 @@ abstract class Puppet : EventEmitter {
     private val HEARTBEAT_COUNTER = AtomicLong()
     private val HOSTIE_KEEPALIVE_TIMEOUT = 15 * 1000L
     private val DEFAULT_WATCHDOG_TIMEOUT = 60L
-    private var memory: MemoryCard
+//    private var memory: MemoryCard
 
     private val executorService = Executors.newSingleThreadScheduledExecutor()
 
@@ -67,8 +66,8 @@ abstract class Puppet : EventEmitter {
         count.addAndGet(1)
         this.puppetOptions = puppetOptions
 
-        this.memory = MemoryCard()
-        this.memory.load()
+//        this.memory = MemoryCard()
+//        this.memory.load()
 
 
         val timeOut = puppetOptions.timeout ?: DEFAULT_WATCHDOG_TIMEOUT
@@ -906,9 +905,9 @@ abstract class Puppet : EventEmitter {
         }
     }
 
-    fun setMemory(memoryCard: MemoryCard){
-        this.memory = memoryCard
-    }
+//    fun setMemory(memoryCard: MemoryCard){
+//        this.memory = memoryCard
+//    }
 
 //    fun getEventBus(): EventBus {
 //        return eb
